requirements.txt
appdirs==1.4.3
apturl==0.5.2
asgiref==3.7.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
cachecontrol==0.13.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
click==7.1.2
cmake==3.18.4.post1
colorama==0.4.3
command-not-found==0.3
cryptography==41.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.2.3
django-restframework==0.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dlib==19.24.2
drf-yasg==1.21.6
duplicity==0.8.12.0
entrypoints==0.3
face-recognition==1.3.0
face-recognition-models==0.3.0
fasteners==0.14.1
filelock==3.0.12
firebase-admin==6.2.0
future==0.18.2
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-status==1.56.2
httplib2==0.22.0
idna==3.4
importlib-metadata==1.5.0
inflection==0.5.1
joblib==1.3.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.5
netifaces==0.10.4
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
packaging==23.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==10.0.0
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.7.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.1.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian==0.1.36+ubuntu1.1
python3-openid==3.2.0
pytz==2023.3
pyxdg==0.26
PyYAML==6.0
reportlab==3.5.34
requests==2.31.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
systemd-python==234
threadpoolctl==3.1.0
typing-extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==2.0.3
usb-creator==0.3.7
virtualenv==20.0.17
wadllib==1.3.3
xkit==0.0.0
zipp==1.0.0
